Chuyển database sang MySQL RDS và cấu hình tự động chạy Migration

+++ WebPetThucTap/Models/ExpCalculator.cs
@@ -2,23 +2,31 @@
 {
     public static class ExpCalculator
     {
-        public static (int Level, string Title, int CurrentExp, int NextExp) Calculate(int totalExp, ExpConfig config)
+        // Đổi CurrentExp và NextExp trong Tuple trả về thành long.
+        // Đổi tham số đầu vào totalExp thành long.
+        public static (int Level, string Title, long CurrentExp, long NextExp) Calculate(long totalExp, ExpConfig config)
         {
-            int level = 1;
-            int expNeededForCurrentLevel = config.BaseExpNeeded;
-            int accumulatedExp = 0;
+            int level = 1; // Cấp độ vẫn giữ int vì không bao giờ vượt quá 2 tỷ
+
+            // Các biến tính toán EXP chuyển hết sang long
+            long expNeededForCurrentLevel = config.BaseExpNeeded;
+            long accumulatedExp = 0;
 
             while (totalExp >= accumulatedExp + expNeededForCurrentLevel)
             {
                 accumulatedExp += expNeededForCurrentLevel;
                 level++;
-                expNeededForCurrentLevel = (int)(expNeededForCurrentLevel * config.Multiplier);
+
+                // Ép kiểu về (long) thay vì (int) để tránh tràn số sau khi nhân hệ số
+                expNeededForCurrentLevel = (long)(expNeededForCurrentLevel * config.Multiplier);
             }
 
-            // Tính số EXP lẻ đang có trong cấp độ hiện tại
-            int expIntoCurrentLevel = totalExp - accumulatedExp;
+            // Số EXP dư ở cấp hiện tại cũng là một con số rất lớn nên dùng long
+            long expIntoCurrentLevel = totalExp - accumulatedExp;
 
             string[] titles = config.TitlesString.Split(',');
+
+            // Các biến dùng làm index mảng và phép toán nhỏ vẫn dùng int
             int titleIndex = (level - 1) / config.LevelsPerTitle;
             int subLevel = (level - 1) % config.LevelsPerTitle;
 
@@ -34,7 +42,7 @@
                 currentTitle += $" {subLevel}";
             }
 
-            // Trả về thêm CurrentExp và NextExp để làm thanh Progress Bar
+            // Trả về tuple đã cập nhật kiểu dữ liệu
             return (level, currentTitle, expIntoCurrentLevel, expNeededForCurrentLevel);
         }
     }
