+++ WebPetThucTap/Models/UserAccount.cs
@@ -17,7 +17,7 @@ namespace WebPetThucTap.Models
         public UserRole Role { get; set; } = UserRole.User;
 
         // Hệ thống EXP
-        public int TotalExp { get; set; } = 0;
+        public long TotalExp { get; set; } = 0;
 
         // Quản lý BAN người dùng
         public DateTime? BannedUntil { get; set; } // Nếu NULL là không bị BAN, nếu có ngày là bị BAN đến ngày đó
